homeworks/shopping/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
tensorflow==2.19.0
shap==0.43.0
imbalanced-learn==0.11.0
python-dotenv==1.0.0
pydantic==2.5.2
psycopg2-binary==2.9.9 